Test for branching task.
I've made a small but noticeable change in the frog level that should be easily verified after a successful build. I need to test whether the compiled executable from TCI is deployed onto the GitHub repository or not. If it is, then I can move on to the next task. If not, then I need to work on the deployment stage.